[#2037] Split into two files
